Port vsock-rs to use rustix
vsock-rs currently uses a mix of nix and libc to handle
the underlying system calls. However, in addition to adding
a few extra dependencies, this means that there is a significant
amount of unneeded unsafe code in the project.

By porting to "rustix", we gain the following advantages:

- Most of the code in the project can be rewritten completely
  safely.
- We avoid going through libc. In other projects this has been
  benchmarked to lead to a 2.5% performance increase.
- On Linux, we reduce the number of dependencies the project
  pulls in. On other platforms, it stays the same.

Blocked on https://github.com/bytecodealliance/rustix/pull/1558

Signed-off-by: John Nunley <dev@notgull.net>